lib: move internal module to internal
