+++ lib/internal/http/client.js
@@ -43,9 +43,9 @@ const {
   HTTPParser,
   isLenient,
   prepareError,
-} = require('_http_common');
-const { OutgoingMessage } = require('_http_outgoing');
-const Agent = require('_http_agent');
+} = require('internal/http/common');
+const { OutgoingMessage } = require('internal/http/outgoing');
+const Agent = require('internal/http/agent');
 const { Buffer } = require('buffer');
 const { defaultTriggerAsyncIdScope } = require('internal/async_hooks');
 const { URL, urlToOptions, searchParamsSymbol } = require('internal/url');
