+++ lib/internal/http/common.js
@@ -32,7 +32,7 @@ const { getOptionValue } = require('internal/options');
 const insecureHTTPParser = getOptionValue('--insecure-http-parser');
 
 const FreeList = require('internal/freelist');
-const incoming = require('_http_incoming');
+const incoming = require('internal/http/incoming');
 const {
   IncomingMessage,
   readStart,
