+++ lib/internal/http/outgoing.js
@@ -38,7 +38,7 @@ const Stream = require('stream');
 const internalUtil = require('internal/util');
 const { kOutHeaders, utcDate, kNeedDrain } = require('internal/http');
 const { Buffer } = require('buffer');
-const common = require('_http_common');
+const common = require('internal/http/common');
 const checkIsHttpToken = common._checkIsHttpToken;
 const checkInvalidHeaderChar = common._checkInvalidHeaderChar;
 const {
