+++ lib/internal/http/server.js
@@ -44,8 +44,8 @@ const {
   isLenient,
   _checkInvalidHeaderChar: checkInvalidHeaderChar,
   prepareError,
-} = require('_http_common');
-const { OutgoingMessage } = require('_http_outgoing');
+} = require('internal/http/common');
+const { OutgoingMessage } = require('internal/http/outgoing');
 const {
   kOutHeaders,
   kNeedDrain,
@@ -55,7 +55,7 @@ const {
   defaultTriggerAsyncIdScope,
   getOrSetAsyncId
 } = require('internal/async_hooks');
-const { IncomingMessage } = require('_http_incoming');
+const { IncomingMessage } = require('internal/http/incoming');
 const {
   connResetException,
   codes
