+++ lib/internal/streams/async_iterator.js
@@ -151,7 +151,7 @@ const createReadableStreamAsyncIterator = (stream) => {
     // v1 stream
 
     if (!Readable) {
-      Readable = require('_stream_readable');
+      Readable = require('internal/streams/readable');
     }
 
     const src = stream;
