+++ lib/internal/streams/duplex.js
@@ -35,8 +35,8 @@ const {
 
 module.exports = Duplex;
 
-const Readable = require('_stream_readable');
-const Writable = require('_stream_writable');
+const Readable = require('internal/streams/readable');
+const Writable = require('internal/streams/writable');
 
 ObjectSetPrototypeOf(Duplex.prototype, Readable.prototype);
 ObjectSetPrototypeOf(Duplex, Readable);
