+++ lib/internal/streams/passthrough.js
@@ -31,7 +31,7 @@ const {
 
 module.exports = PassThrough;
 
-const Transform = require('_stream_transform');
+const Transform = require('internal/streams/transform');
 ObjectSetPrototypeOf(PassThrough.prototype, Transform.prototype);
 ObjectSetPrototypeOf(PassThrough, Transform);
