+++ lib/internal/streams/pipeline.js
@@ -212,7 +212,7 @@ function pipeline(...streams) {
         }
       } else {
         if (!PassThrough) {
-          PassThrough = require('_stream_passthrough');
+          PassThrough = require('internal/streams/passthrough');
         }
 
         // If the last argument to pipeline is not a stream
