+++ lib/internal/streams/transform.js
@@ -72,7 +72,7 @@ module.exports = Transform;
 const {
   ERR_METHOD_NOT_IMPLEMENTED
 } = require('internal/errors').codes;
-const Duplex = require('_stream_duplex');
+const Duplex = require('internal/streams/duplex');
 ObjectSetPrototypeOf(Transform.prototype, Duplex.prototype);
 ObjectSetPrototypeOf(Transform, Duplex);
