+++ lib/internal/streams/writable.js
