+++ lib/internal/tls/wrap.js
@@ -43,7 +43,7 @@ const crypto = require('crypto');
 const EE = require('events');
 const net = require('net');
 const tls = require('tls');
-const common = require('_tls_common');
+const common = require('internal/tls/common');
 const JSStreamSocket = require('internal/js_stream_socket');
 const { Buffer } = require('buffer');
 const debug = require('internal/util/debuglog').debuglog('tls');
