+++ lib/stream.js
@@ -29,11 +29,11 @@ const internalBuffer = require('internal/buffer');
 // to avoid a cross-reference(require) issues
 const Stream = module.exports = require('internal/streams/legacy');
 
-Stream.Readable = require('_stream_readable');
-Stream.Writable = require('_stream_writable');
-Stream.Duplex = require('_stream_duplex');
-Stream.Transform = require('_stream_transform');
-Stream.PassThrough = require('_stream_passthrough');
+Stream.Readable = require('internal/streams/readable');
+Stream.Writable = require('internal/streams/writable');
+Stream.Duplex = require('internal/streams/duplex');
+Stream.Transform = require('internal/streams/transform');
+Stream.PassThrough = require('internal/streams/passthrough');
 
 Stream.pipeline = pipeline;
 Stream.finished = eos;
