+++ lib/zlib.js
@@ -45,7 +45,7 @@ const {
   },
   hideStackFrames
 } = require('internal/errors');
-const Transform = require('_stream_transform');
+const Transform = require('internal/streams/transform');
 const {
   deprecate
 } = require('internal/util');
