+++ node.gyp
@@ -141,6 +141,12 @@
       'lib/internal/fs/utils.js',
       'lib/internal/fs/watchers.js',
       'lib/internal/http.js',
+      'lib/internal/http/agent.js',
+      'lib/internal/http/client.js',
+      'lib/internal/http/common.js',
+      'lib/internal/http/incoming.js',
+      'lib/internal/http/outgoing.js',
+      'lib/internal/http/server.js',
       'lib/internal/heap_utils.js',
       'lib/internal/histogram.js',
       'lib/internal/idna.js',
@@ -197,6 +203,8 @@
       'lib/internal/test/binding.js',
       'lib/internal/timers.js',
       'lib/internal/tls.js',
+      'lib/internal/tls/common.js',
+      'lib/internal/tls/wrap.js',
       'lib/internal/trace_events_async_hooks.js',
       'lib/internal/tty.js',
       'lib/internal/url.js',
@@ -227,6 +235,11 @@
       'lib/internal/streams/state.js',
       'lib/internal/streams/pipeline.js',
       'lib/internal/streams/end-of-stream.js',
+      'lib/internal/streams/readable.js',
+      'lib/internal/streams/writable.js',
+      'lib/internal/streams/duplex.js',
+      'lib/internal/streams/transform.js',
+      'lib/internal/streams/passthrough.js',
       'deps/v8/tools/splaytree.js',
       'deps/v8/tools/codemap.js',
       'deps/v8/tools/consarray.js',
