+++ test/parallel/test-stream2-base64-single-char-read-end.js
@@ -21,8 +21,7 @@
 
 'use strict';
 require('../common');
-const R = require('_stream_readable');
-const W = require('_stream_writable');
+const { Readable: R, Writable: W } = require('stream');
 const assert = require('assert');
 
 const src = new R({ encoding: 'base64' });
