+++ test/parallel/test-stream2-compatibility.js
@@ -21,8 +21,7 @@
 
 'use strict';
 require('../common');
-const R = require('_stream_readable');
-const W = require('_stream_writable');
+const { Readable: R, Writable: W } = require('stream');
 const assert = require('assert');
 
 let ondataCalled = 0;
