+++ test/parallel/test-stream2-readable-from-list.js
@@ -23,7 +23,7 @@
 'use strict';
 require('../common');
 const assert = require('assert');
-const fromList = require('_stream_readable')._fromList;
+const { Readable: { _fromList: fromList } } = require('stream');
 const BufferList = require('internal/streams/buffer_list');
 const util = require('util');
